Gives the user the power to decide when to apply an box update.
Changes the update progress to ask the user to apply an available update until the update is applied (for every time the user calls 'vagrant up'). So far the vagrant-box-updater just asked once and if the user denies updating, the updater keeps track of this decision and ignores this update until another update is available. The auto-update behaviour is not affected by this commit.